Fix a bug in Tracker
Issue:
There are many repeated annotations. As a result, we see bounding boxes' labels overlap with many different numbers. After Tracker.track(), the number of objects increase and are repeated.

Root cause:
When we track objects to identify whether they are the same objects, the logic of the two for loops is incorrect.

Solution:
- We should make updating objects the first for loop,
- then we compare one updating object with all base objects in the second for loop.
- If we found that the updating object is the same object as one of the base object, we break the second for loop.
- Otherwise, after the second for loop ends, we give the updating object a different object ID. Then, we proceed to the next updating object.